feat: add slices to keys implementations
Prior to this commit, we provided implementations only for `Vec`.
However, the exact same implementation can be easily extended to slices,
as both types are almost transparent.

This commit introduces such implementation. It will allow the user to
query a radix tree that has a key of integers as key, using a slice of
integers. That is possible because the get fucntion will accept anything
that the key borrows to, and vec borrows to slice natively.